pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "svs-label-ocr"
version = "0.3.2"
description = "Batch OCR for SVS label associated images."
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.0",
    "openai>=1.0.0",
    "openslide-python>=1.3.1",
    "pillow>=11.0.0",
    "pytesseract>=0.3.10",
]

[project.scripts]
svs-label-ocr = "svs_label_ocr.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
